+++ CHANGELOG.md
@@ -16,6 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - DscResource.Base
   - Move code coverage task in the pipeline to use task `PublishCodeCoverageResults@2`.
+- `Get-LocalizedDataRecursive`
+  - Move strings to localized versions ([issue #7](https://github.com/dsccommunity/DscResource.Base/issues/7)).
+  - Fix various formatting issues
 
 ## [1.1.0] - 2023-02-26